Change names (notes to self in commit msg)
OK, so basically afaict, the ~ONLY~ things Tabula Rasa does are:

- Set up propgroups to define what can be edited via the wizard.
- Print default CSS.

THAT'S IT. The actual page structure all comes 100% from core2. This is in
contrast to some other layouts that override the page write functions. (Heck,
even my theme overrides Page::print_header.) So, IF it makes sense to replace
the floats-based layout with a flexbox thing, and IF the current page markup
structure is granular enough to permit it, replacing Rasa ought to be hella
easy, because Rasa basically does nothing. It only gets hard if there's actual
markup changes to make.