FIX EVERYTHING, except...... it's ONLY for this year's firefox. D:<
(I even got the tertiary module space to go to the bottom and behave!!!)

The key is row-gap and column-gap as applied to flex layouts. It's in the
"CSS box alignment module level 3" w3c working draft, and so far Fixefox (in the
60s) is the only thing that supports that. So it's beautiful for me! And kind of
crap for everyone else.

This seems like a good spot to leave off on this. I'll come back to it in a year
or two when the browser support is better.