Experimental awful hack: different colors for foreign entry pages
I want to be able to distinguish between my entries and ones belonging to other
people, WITHOUT having to use the gnarly site skin.

EDIT: This basically doesn't work. I tricked myself.

Well... I mean, it works as long as you're logged in as me, or logged out. But
if you're logged in as SOMEONE ELSE, then my entry pages won't have
`.my-journal`, so it'll end up looking weird and confusing. (Assuming you don't
have it set to display them in your style in the first place, but anyway.)

I think this might be impossible, with the HTML classes I get from core2. It was
an interesting experiment, but the amount of logic I'd have to override is
insane, and I still don't have any guarantee I could get it working the way I
want.